+++ test/list.js
@@ -1,5 +1,5 @@
-var assert = require('assert');
-var List = require('../lib').List;
+const assert = require('assert');
+const List = require('../lib').List;
 
 function getFirstArg(data) {
     return data;
@@ -11,17 +11,18 @@ function getData(item) {
 
 function getIteratorState(data, item, list) {
     return {
-        context: this,
-        data: data,
-        item: item,
-        list: list
+        thisArg: this,
+        data,
+        item,
+        list
     };
 }
 
 function createIteratorStateCollector(retValue) {
-    var states = [];
+    const states = [];
+
     return {
-        states: states,
+        states,
         collect: function() {
             states.push(getIteratorState.apply(this, arguments));
             return retValue;
@@ -30,16 +31,18 @@ function createIteratorStateCollector(retValue) {
 }
 
 function toArray(list) {
-    var cursor = list.head;
-    var prev = null;
+    let cursor = list.head;
+    let prev = null;
 
     while (cursor !== null) {
         if (cursor.prev !== prev) {
             throw new Error('Wrong prev reference');
         }
+
         if (prev !== null && prev.next !== cursor) {
             throw new Error('Wrong next reference');
         }
+
         prev = cursor;
         cursor = cursor.next;
     }
@@ -56,41 +59,41 @@ function toArray(list) {
 }
 
 function createIteratorTests(list, method, items, until, retValue) {
-    var iterate = until !== undefined
+    const iterate = until !== undefined
         ? list[method].bind(list, until)
         : list[method].bind(list);
 
     it('iterate', function() {
-        var collector = createIteratorStateCollector(retValue);
+        const collector = createIteratorStateCollector(retValue);
 
         iterate(collector.collect);
         assert.deepEqual(collector.states, items.map(function(item) {
             return {
-                context: list,
-                list: list,
-                item: item,
+                thisArg: list,
+                list,
+                item,
                 data: item.data
             };
         }));
     });
 
-    it('iterate with context', function() {
-        var context = {};
-        var collector = createIteratorStateCollector(retValue);
+    it('iterate with thisArg', function() {
+        const thisArg = {};
+        const collector = createIteratorStateCollector(retValue);
 
-        iterate(collector.collect, context);
+        iterate(collector.collect, thisArg);
         assert.deepEqual(collector.states, items.map(function(item) {
             return {
-                context: context,
-                list: list,
-                item: item,
+                thisArg: thisArg,
+                list,
+                item,
                 data: item.data
             };
         }));
     });
 
     it('nested iterate', function() {
-        var collector = createIteratorStateCollector(retValue);
+        const collector = createIteratorStateCollector(retValue);
 
         iterate(function() {
             iterate(collector.collect);
@@ -100,9 +103,9 @@ function createIteratorTests(list, method, items, until, retValue) {
         assert.deepEqual(collector.states, items.reduce(function(expected) {
             return expected.concat(items.map(function(item) {
                 return {
-                    context: list,
-                    list: list,
-                    item: item,
+                    thisArg: list,
+                    list,
+                    item,
                     data: item.data
                 };
             }));
@@ -111,28 +114,29 @@ function createIteratorTests(list, method, items, until, retValue) {
 }
 
 function createIteratorWithModificationTests(list, method, items, until, retValue) {
-    var iterate = until !== undefined
-        ? function(list) {
-            var untilCopy = null;
-            var cursor = list.head;
+    const iterate = until !== undefined
+        ? function(list, ...args) {
+            let untilCopy = null;
+            let cursor = list.head;
 
             while (cursor) {
                 if (cursor.data === until.data) {
                     untilCopy = cursor;
                     break;
                 }
+
                 cursor = cursor.next;
             }
 
-            list[method].apply(list, [untilCopy].concat(Array.prototype.slice.call(arguments, 1)));
+            list[method].apply(list, [untilCopy, ...args]);
         }
-        : function(list) {
-            list[method].apply(list, Array.prototype.slice.call(arguments, 1));
+        : function(list, ...args) {
+            list[method].apply(list, args);
         };
 
     it('remove items on iterate', function() {
-        var iterateList = list.copy();
-        var removed = [];
+        const iterateList = list.copy();
+        const removed = [];
 
         iterate(iterateList, function(data, item) {
             removed.push(data);
@@ -145,15 +149,15 @@ function createIteratorWithModificationTests(list, method, items, until, retValu
     });
 
     it('insert items on iterate', function() {
-        var iterateList = list.copy();
-        var inserted = [];
-        var order = [];
-        var idx = 0;
+        const iterateList = list.copy();
+        const inserted = [];
+        const order = [];
+        let idx = 0;
 
         iterate(iterateList, function(data, item) {
             if (!/^ins\d+$/.test(data)) {
-                var before = List.createItem('ins' + idx);
-                var after = List.createItem('ins' + idx);
+                const before = List.createItem('ins' + idx);
+                const after = List.createItem('ins' + idx);
 
                 iterateList.insert(before, item);
                 iterateList.insert(after, item.next);
@@ -169,19 +173,19 @@ function createIteratorWithModificationTests(list, method, items, until, retValu
             return res.concat(item.data, 'ins' + idx);
         }, []));
 
-        assert.deepEqual(iterateList.map(getFirstArg).toArray(), list.toArray().reduce(function(res, data, idx) {
-            var idx = order.indexOf(data);
+        assert.deepEqual(iterateList.map(getFirstArg).toArray(), list.toArray().reduce(function(res, data) {
+            const idx = order.indexOf(data);
             return res.concat('ins' + idx, data, 'ins' + idx);
         }, []));
     });
 }
 
 describe('List', function() {
-    var foo = {};
-    var bar = {};
-    var empty;
-    var list1;
-    var list2;
+    const foo = {};
+    const bar = {};
+    let empty;
+    let list1;
+    let list2;
 
     beforeEach(function() {
         empty = new List();
@@ -205,14 +209,14 @@ describe('List', function() {
         });
     });
 
-    it('#getSize()', function() {
-        assert.equal(empty.getSize(), 0);
-        assert.equal(list1.getSize(), 1);
-        assert.equal(list2.getSize(), 2);
+    it('#size', function() {
+        assert.equal(empty.size, 0);
+        assert.equal(list1.size, 1);
+        assert.equal(list2.size, 2);
     });
 
     it('#fromArray()', function() {
-        var list = new List().fromArray([foo, bar]);
+        const list = new List().fromArray([foo, bar]);
 
         assert.equal(list.head.data, foo);
         assert.equal(list.tail.data, bar);
@@ -230,45 +234,39 @@ describe('List', function() {
         assert.deepEqual(list2.toJSON(), [foo, bar]);
     });
 
-    it('#isEmpty()', function() {
-        assert.equal(empty.isEmpty(), true);
-        assert.equal(list1.isEmpty(), false);
-        assert.equal(list2.isEmpty(), false);
+    it('#isEmpty', function() {
+        assert.equal(empty.isEmpty, true);
+        assert.equal(list1.isEmpty, false);
+        assert.equal(list2.isEmpty, false);
     });
 
-    it('#first()', function() {
-        assert.equal(empty.first(), null);
-        assert.equal(list1.first(), foo);
-        assert.equal(list2.first(), foo);
+    it('#first', function() {
+        assert.equal(empty.first, null);
+        assert.equal(list1.first, foo);
+        assert.equal(list2.first, foo);
     });
 
-    it('#last()', function() {
-        assert.equal(empty.last(), null);
-        assert.equal(list1.last(), foo);
-        assert.equal(list2.last(), bar);
-    });
-
-    describe('#each()', function() {
-        var iterateList = new List().fromArray([foo, bar]);
-        createIteratorTests(iterateList, 'each', [iterateList.head, iterateList.tail]);
-        createIteratorWithModificationTests(iterateList, 'each', [iterateList.head, iterateList.tail]);
+    it('#last', function() {
+        assert.equal(empty.last, null);
+        assert.equal(list1.last, foo);
+        assert.equal(list2.last, bar);
     });
 
     describe('#forEach()', function() {
-        var iterateList = new List().fromArray([foo, bar]);
+        const iterateList = new List().fromArray([foo, bar]);
         createIteratorTests(iterateList, 'forEach', [iterateList.head, iterateList.tail]);
         createIteratorWithModificationTests(iterateList, 'forEach', [iterateList.head, iterateList.tail]);
     });
 
-    describe('#eachRight()', function() {
-        var iterateList = new List().fromArray([foo, bar]);
-        createIteratorTests(iterateList, 'eachRight', [iterateList.tail, iterateList.head]);
-        createIteratorWithModificationTests(iterateList, 'eachRight', [iterateList.tail, iterateList.head]);
+    describe('#forEachRight()', function() {
+        const iterateList = new List().fromArray([foo, bar]);
+        createIteratorTests(iterateList, 'forEachRight', [iterateList.tail, iterateList.head]);
+        createIteratorWithModificationTests(iterateList, 'forEachRight', [iterateList.tail, iterateList.head]);
     });
 
     describe('#nextUntil()', function() {
         it('should not iterate when start is null', function() {
-            var count = 0;
+            let count = 0;
 
             list2.nextUntil(null, function() {
                 count++;
@@ -278,7 +276,7 @@ describe('List', function() {
         });
 
         it('should stop iterate when callback returns true', function() {
-            var count = 0;
+            let count = 0;
 
             list2.nextUntil(list2.head, function() {
                 count++;
@@ -288,14 +286,14 @@ describe('List', function() {
             assert.equal(count, 1);
         });
 
-        var iterateList = new List().fromArray([foo, bar]);
+        const iterateList = new List().fromArray([foo, bar]);
         createIteratorTests(iterateList, 'nextUntil', [iterateList.head, iterateList.tail], iterateList.head);
         createIteratorWithModificationTests(iterateList, 'nextUntil', [iterateList.head, iterateList.tail], iterateList.head);
     });
 
     describe('#prevUntil()', function() {
         it('should not iterate when start is null', function() {
-            var count = 0;
+            let count = 0;
 
             list2.prevUntil(null, function() {
                 count++;
@@ -305,7 +303,7 @@ describe('List', function() {
         });
 
         it('should stop iterate when callback returns true', function() {
-            var count = 0;
+            let count = 0;
 
             list2.prevUntil(list2.tail, function() {
                 count++;
@@ -315,7 +313,7 @@ describe('List', function() {
             assert.equal(count, 1);
         });
 
-        var iterateList = new List().fromArray([foo, bar]);
+        const iterateList = new List().fromArray([foo, bar]);
         createIteratorTests(iterateList, 'prevUntil', [iterateList.tail, iterateList.head], iterateList.tail);
         createIteratorWithModificationTests(iterateList, 'prevUntil', [iterateList.tail, iterateList.head], iterateList.tail);
     });
@@ -330,7 +328,7 @@ describe('List', function() {
         });
 
         it('should stop on first match', function() {
-            var count = 0;
+            let count = 0;
             assert.equal(list2.some(function(data) {
                 count++;
                 return data === foo;
@@ -338,7 +336,7 @@ describe('List', function() {
             assert.equal(count, 1);
         });
 
-        var iterateList = new List().fromArray([foo, bar]);
+        const iterateList = new List().fromArray([foo, bar]);
         createIteratorTests(iterateList, 'some', [iterateList.head, iterateList.tail]);
     });
 
@@ -347,16 +345,16 @@ describe('List', function() {
             assert.deepEqual(empty.map(getFirstArg).toArray(), []);
             assert.deepEqual(list1.map(getFirstArg).toArray(), [foo]);
 
-            var foo2 = {};
-            var bar2 = {};
-            var mapped = list2.map(function(node) {
+            const foo2 = {};
+            const bar2 = {};
+            const mapped = list2.map(function(node) {
                 return node === foo ? foo2 : bar2;
             });
             assert.deepEqual(mapped.toArray(), [foo2, bar2]);
             assert.deepEqual(list2.toArray(), [foo, bar]);
         });
 
-        var iterateList = new List().fromArray([foo, bar]);
+        const iterateList = new List().fromArray([foo, bar]);
         createIteratorTests(iterateList, 'map', [iterateList.head, iterateList.tail]);
     });
 
@@ -365,14 +363,14 @@ describe('List', function() {
             assert.deepEqual(empty.filter(getFirstArg).toArray(), []);
             assert.deepEqual(list2.filter(Boolean).toArray(), [foo, bar]);
 
-            var filtered = list2.filter(function(node) {
+            const filtered = list2.filter(function(node) {
                 return node === bar;
             });
             assert.deepEqual(filtered.toArray(), [bar]);
             assert.deepEqual(list2.toArray(), [foo, bar]);
         });
 
-        var iterateList = new List().fromArray([foo, bar]);
+        const iterateList = new List().fromArray([foo, bar]);
         createIteratorTests(iterateList, 'filter', [iterateList.head, iterateList.tail]);
     });
 
@@ -395,7 +393,7 @@ describe('List', function() {
     });
 
     it('#copy()', function() {
-        var copy = list2.copy();
+        const copy = list2.copy();
 
         assert.notEqual(copy, list2);
         assert.notEqual(copy.head, list2.head);
@@ -405,7 +403,7 @@ describe('List', function() {
     });
 
     it('#prepend()', function() {
-        var qux = {};
+        const qux = {};
 
         assert.deepEqual(toArray(empty.prepend(List.createItem(qux))), [qux]);
         assert.deepEqual(toArray(list1.prepend(List.createItem(qux))), [qux, foo]);
@@ -414,7 +412,7 @@ describe('List', function() {
     });
 
     it('#prependData()', function() {
-        var qux = {};
+        const qux = {};
 
         assert.deepEqual(toArray(empty.prependData(qux)), [qux]);
         assert.deepEqual(toArray(list1.prependData(qux)), [qux, foo]);
@@ -423,7 +421,7 @@ describe('List', function() {
     });
 
     it('#unshift()', function() {
-        var qux = {};
+        const qux = {};
 
         empty.unshift(qux);
         assert.deepEqual(empty.toArray(), [qux]);
@@ -439,8 +437,8 @@ describe('List', function() {
 
     describe('#shift()', function() {
         it('should remove first item', function() {
-            var head = list2.head;
-            var tail = list2.tail;
+            const head = list2.head;
+            const tail = list2.tail;
 
             assert.equal(list2.shift(), head);
             assert.equal(list2.shift(), tail);
@@ -453,7 +451,7 @@ describe('List', function() {
     });
 
     it('#append()', function() {
-        var qux = {};
+        const qux = {};
 
         assert.deepEqual(toArray(empty.append(List.createItem(qux))), [qux]);
         assert.deepEqual(toArray(list1.append(List.createItem(qux))), [foo, qux]);
@@ -462,7 +460,7 @@ describe('List', function() {
     });
 
     it('#appendData()', function() {
-        var qux = {};
+        const qux = {};
 
         assert.deepEqual(toArray(empty.appendData(qux)), [qux]);
         assert.deepEqual(toArray(list1.appendData(qux)), [foo, qux]);
@@ -471,7 +469,7 @@ describe('List', function() {
     });
 
     it('#push()', function() {
-        var qux = {};
+        const qux = {};
 
         empty.push(qux);
         assert.deepEqual(empty.toArray(), [qux]);
@@ -487,8 +485,8 @@ describe('List', function() {
 
     describe('#pop()', function() {
         it('should remove last item', function() {
-            var head = list2.head;
-            var tail = list2.tail;
+            const head = list2.head;
+            const tail = list2.tail;
 
             assert.equal(list2.pop(), tail);
             assert.equal(list2.pop(), head);
@@ -502,27 +500,27 @@ describe('List', function() {
 
     describe('#insert()', function() {
         it('should append when no ref item', function() {
-            var afterTail = {};
-            var res = list2.insert(List.createItem(afterTail));
+            const afterTail = {};
+            const res = list2.insert(List.createItem(afterTail));
 
             assert.equal(res, list2);
             assert.equal(list2.tail.data, afterTail);
         });
 
         it('should insert before ref item', function() {
-            var beforeHead = {};
-            var res = list2.insert(List.createItem(beforeHead), list2.head);
+            const beforeHead = {};
+            const res = list2.insert(List.createItem(beforeHead), list2.head);
 
             assert.equal(res, list2);
             assert.equal(list2.head.data, beforeHead);
         });
 
         it('insert in the middle', function() {
-            var qux = {};
-            var inserted = List.createItem(qux);
-            var list2head = list2.head;
-            var list2tail = list2.tail;
-            var res = list2.insert(inserted, list2.tail);
+            const qux = {};
+            const inserted = List.createItem(qux);
+            const list2head = list2.head;
+            const list2tail = list2.tail;
+            const res = list2.insert(inserted, list2.tail);
 
             assert.equal(res, list2);
             assert.deepEqual(list2head, {
@@ -543,7 +541,7 @@ describe('List', function() {
         });
 
         it('insert the item before an item that doesn\'t belong to list', function() {
-            var inserted = List.createItem({});
+            const inserted = List.createItem({});
 
             assert.throws(function() {
                 list2.insert(inserted, list1.head);
@@ -553,26 +551,26 @@ describe('List', function() {
 
     describe('#insertData()', function() {
         it('should append when no ref item', function() {
-            var afterTail = {};
-            var res = list2.insertData(afterTail);
+            const afterTail = {};
+            const res = list2.insertData(afterTail);
 
             assert.equal(res, list2);
             assert.equal(list2.tail.data, afterTail);
         });
 
         it('should insert before ref item', function() {
-            var beforeHead = {};
-            var res = list2.insertData(beforeHead, list2.head);
+            const beforeHead = {};
+            const res = list2.insertData(beforeHead, list2.head);
 
             assert.equal(res, list2);
             assert.equal(list2.head.data, beforeHead);
         });
 
         it('insert in the middle', function() {
-            var qux = {};
-            var list2head = list2.head;
-            var list2tail = list2.tail;
-            var res = list2.insertData(qux, list2.tail);
+            const qux = {};
+            const list2head = list2.head;
+            const list2tail = list2.tail;
+            const res = list2.insertData(qux, list2.tail);
 
             assert.equal(res, list2);
             assert.deepEqual(list2head, {
@@ -601,8 +599,8 @@ describe('List', function() {
 
     describe('#remove()', function() {
         it('clear a list', function() {
-            var head = list2.head;
-            var tail = list2.tail;
+            const head = list2.head;
+            const tail = list2.tail;
 
             list2.remove(list2.tail);
             list2.remove(list2.head);
@@ -622,8 +620,8 @@ describe('List', function() {
         });
 
         it('clear a list in reverse order', function() {
-            var head = list2.head;
-            var tail = list2.tail;
+            const head = list2.head;
+            const tail = list2.tail;
 
             list2.remove(list2.head);
             list2.remove(list2.tail);
@@ -657,9 +655,9 @@ describe('List', function() {
 
     describe('#prependList()', function() {
         it('prepend non-empty list to non-empty list', function() {
-            var list1head = list1.head;
-            var list2head = list2.head;
-            var res = list2.prependList(list1);
+            const list1head = list1.head;
+            const list2head = list2.head;
+            const res = list2.prependList(list1);
 
             assert.equal(res, list2);
             assert.equal(list1.head, null);
@@ -677,9 +675,9 @@ describe('List', function() {
         });
 
         it('prepend non-empty list to empty list', function() {
-            var list2head = list2.head;
-            var list2tail = list2.tail;
-            var res = empty.prependList(list2);
+            const list2head = list2.head;
+            const list2tail = list2.tail;
+            const res = empty.prependList(list2);
 
             assert.equal(res, empty);
             assert.equal(empty.head, list2head);
@@ -689,8 +687,8 @@ describe('List', function() {
         });
 
         it('prepend empty list to non-empty', function() {
-            var list2head = list2.head;
-            var res = list2.prependList(empty);
+            const list2head = list2.head;
+            const res = list2.prependList(empty);
 
             assert.equal(res, list2);
             assert.equal(empty.head, null);
@@ -700,9 +698,9 @@ describe('List', function() {
 
     describe('#appendList()', function() {
         it('append non-empty list to non-empty list', function() {
-            var list1head = list1.head;
-            var list2tail = list2.tail;
-            var res = list2.appendList(list1);
+            const list1head = list1.head;
+            const list2tail = list2.tail;
+            const res = list2.appendList(list1);
 
             assert.equal(res, list2);
             assert.equal(list1.head, null);
@@ -720,9 +718,9 @@ describe('List', function() {
         });
 
         it('append non-empty list to empty list', function() {
-            var list2head = list2.head;
-            var list2tail = list2.tail;
-            var res = empty.appendList(list2);
+            const list2head = list2.head;
+            const list2tail = list2.tail;
+            const res = empty.appendList(list2);
 
             assert.equal(res, empty);
             assert.equal(empty.head, list2head);
@@ -732,8 +730,8 @@ describe('List', function() {
         });
 
         it('append empty list to non-empty', function() {
-            var list2tail = list2.tail;
-            var res = list2.appendList(empty);
+            const list2tail = list2.tail;
+            const res = list2.appendList(empty);
 
             assert.equal(res, list2);
             assert.equal(empty.head, null);
@@ -743,8 +741,8 @@ describe('List', function() {
 
     describe('#insertList()', function() {
         it('add non-empty list to non-empty list', function() {
-            var list1head = list1.head;
-            var res = list2.insertList(list1, list2.tail);
+            const list1head = list1.head;
+            const res = list2.insertList(list1, list2.tail);
 
             assert.equal(res, list2);
             assert.equal(list1.head, null);
@@ -767,9 +765,9 @@ describe('List', function() {
         });
 
         it('add non-empty list to empty list', function() {
-            var list2head = list2.head;
-            var list2tail = list2.tail;
-            var res = empty.insertList(list2);
+            const list2head = list2.head;
+            const list2tail = list2.tail;
+            const res = empty.insertList(list2);
 
             assert.equal(res, empty);
             assert.equal(empty.head, list2head);
@@ -779,8 +777,8 @@ describe('List', function() {
         });
 
         it('add empty list to non-empty', function() {
-            var list2tail = list2.tail;
-            var res = list2.insertList(empty);
+            const list2tail = list2.tail;
+            const res = list2.insertList(empty);
 
             assert.equal(res, list2);
             assert.equal(empty.head, null);
@@ -790,7 +788,7 @@ describe('List', function() {
 
     describe('#replace()', function() {
         it('replace for an item', function() {
-            var qux = {};
+            const qux = {};
             list2.replace(list2.tail, List.createItem(qux));
 
             assert.deepEqual(toArray(list2), [foo, qux]);