+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## next
+
+- Removed `List#each()` and `List#eachRight()` methods, `List#forEach()` and `List#forEachRight()` should be used instead
+- Changed `List` to be iterable (iterates data)
+- Changed `List#first`, `List#last` and `List#isEmpty` to getters
+- Changed `List#getSize()` method to `List#size` getters
+
 ## 1.0.0-alpha.37 (October 22, 2019)
 
 - Bumped `source-map` version to `^0.6.1` to fix source map generation inconsistency across node.js versions due to mappings sorting bug and v8 moving to [a stable Array#sort](https://v8.dev/blog/array-sort) ([fix commit](https://github.com/mozilla/source-map/commit/f35a2e4212dd025cb5e1fc219e7ac8a4b96c2cc9) in `source-map`)